+++ src/Felicity/Query.fs
@@ -8,6 +8,8 @@ open Errors
 [<AutoOpen>]
 module private QueryParseHelpers =
 
+  open System.Text.RegularExpressions
+
   let parseBool = function
     | "true" -> Ok true
     | "false" -> Ok false
@@ -26,12 +28,24 @@ module private QueryParseHelpers =
   /// Converts a string conforming to the ISO 8601-1:2019 format to a DateTime.
   let parseDateTime (str: string) =
     try System.Text.Json.JsonSerializer.Deserialize<DateTime> ("\"" + str + "\"") |> Ok
-    with _ -> Error [queryInvalidDateTimeUnnamed str]
+    with _ -> Error [queryInvalidParsedErrMsgUnnamed str "Expected a valid ISO 8601-1:2019 date-time"]
 
   /// Converts a string conforming to the ISO 8601-1:2019 format to a DateTimeOffset.
-  let parseDateTimeOffset (str: string) =
+  let parseDateTimeOffsetAllowMissingOffset (str: string) =
     try System.Text.Json.JsonSerializer.Deserialize<DateTimeOffset> ("\"" + str + "\"") |> Ok
-    with _ -> Error [queryInvalidDateTimeUnnamed str]
+    with _ -> Error [queryInvalidParsedErrMsgUnnamed str "Expected a valid ISO 8601-1:2019 date-time including an offset (e.g. 'Z' or '+01:00')"]
+
+
+  /// Converts a string conforming to the ISO 8601-1:2019 format to a DateTimeOffset.
+  let parseDateTimeOffset =
+    let r = Regex("(?>Z|(?>\+|-)\d\d:\d\d)$", RegexOptions.Compiled)
+    let errMsg = "Expected a valid ISO 8601-1:2019 date-time including an offset (e.g. 'Z' or '+01:00')"
+    fun (str: string) ->
+      try
+        let res = System.Text.Json.JsonSerializer.Deserialize<DateTimeOffset> ("\"" + str + "\"")
+        if r.IsMatch str then Ok res
+        else Error [queryInvalidParsedErrMsgUnnamed str errMsg]
+      with _ -> Error [queryInvalidParsedErrMsgUnnamed str errMsg]
 
 
 
@@ -397,7 +411,7 @@ type Filter =
     Filter.Field(field, parseDateTime)
 
   static member Field(field: FieldQueryParser<'ctx, 'entity, 'attr, DateTimeOffset>) =
-    Filter.Field(field, parseDateTimeOffset)
+    Filter.Field(field, parseDateTimeOffsetAllowMissingOffset)
 
   static member Field(path: Relationship<'ctx, 'entity, 'relatedEntity, 'relatedId>, field: FieldQueryParser<'ctx, 'relatedEntity, 'attr, 'serialized>, toSerialized: string -> Result<'serialized, Error list>) =
     SingleFilter<'ctx, 'attr>(path.Name + "." + field.Name, fun ctx str -> toSerialized str |> Job.result |> JobResult.bind (field.ToDomain ctx))
@@ -421,7 +435,7 @@ type Filter =
     Filter.Field(path, field, parseDateTime)
 
   static member Field(path: Relationship<'ctx, 'entity, 'relatedEntity, 'relatedId>, field: FieldQueryParser<'ctx, 'relatedEntity, 'attr, DateTimeOffset>) =
-    Filter.Field(path, field, parseDateTimeOffset)
+    Filter.Field(path, field, parseDateTimeOffsetAllowMissingOffset)
 
   static member Field(path1: Relationship<'ctx, 'entity, 'relatedEntity1, 'relatedId1>, path2: Relationship<'ctx, 'relatedEntity1, 'relatedEntity2, 'relatedId2>, field: FieldQueryParser<'ctx, 'relatedEntity2, 'attr, 'serialized>, toSerialized: string -> Result<'serialized, Error list>) =
     SingleFilter<'ctx, 'attr>(path1.Name + "." + path2.Name + "." + field.Name, fun ctx str -> toSerialized str |> Job.result |> JobResult.bind (field.ToDomain ctx))
@@ -445,7 +459,7 @@ type Filter =
     Filter.Field(path1, path2, field, parseDateTime)
 
   static member Field(path1: Relationship<'ctx, 'entity, 'relatedEntity1, 'relatedId1>, path2: Relationship<'ctx, 'relatedEntity1, 'relatedEntity2, 'relatedId2>, field: FieldQueryParser<'ctx, 'relatedEntity2, 'attr, DateTimeOffset>) =
-    Filter.Field(path1, path2, field, parseDateTimeOffset)
+    Filter.Field(path1, path2, field, parseDateTimeOffsetAllowMissingOffset)
 
   static member Field(path1: Relationship<'ctx, 'entity, 'relatedEntity1, 'relatedId1>, path2: Relationship<'ctx, 'relatedEntity1, 'relatedEntity2, 'relatedId2>, path3: Relationship<'ctx, 'relatedEntity2, 'relatedEntity3, 'relatedId3>, field: FieldQueryParser<'ctx, 'relatedEntity3, 'attr, 'serialized>, toSerialized: string -> Result<'serialized, Error list>) =
     SingleFilter<'ctx, 'attr>(path1.Name + "." + path2.Name + "." + path3.Name + "." + field.Name, fun ctx str -> toSerialized str |> Job.result |> JobResult.bind (field.ToDomain ctx))
@@ -469,7 +483,7 @@ type Filter =
     Filter.Field(path1, path2, path3, field, parseDateTime)
 
   static member Field(path1: Relationship<'ctx, 'entity, 'relatedEntity1, 'relatedId1>, path2: Relationship<'ctx, 'relatedEntity1, 'relatedEntity2, 'relatedId2>, path3: Relationship<'ctx, 'relatedEntity2, 'relatedEntity3, 'relatedId3>, field: FieldQueryParser<'ctx, 'relatedEntity3, 'attr, DateTimeOffset>) =
-    Filter.Field(path1, path2, path3, field, parseDateTimeOffset)
+    Filter.Field(path1, path2, path3, field, parseDateTimeOffsetAllowMissingOffset)
 
   static member ParsedJobRes(name, parse: 'ctx -> string -> Job<Result<'a, Error list>>) =
     SingleFilter<'ctx, 'a>(name, parse)
@@ -801,6 +815,9 @@ type Query =
 
   static member DateTimeOffset(queryParamName) : CustomQueryParam<'ctx, DateTimeOffset> =
     CustomQueryParam<'ctx, DateTimeOffset>(queryParamName, fun _ -> parseDateTimeOffset >> Job.result)
+
+  static member DateTimeOffsetAllowMissingOffset(queryParamName) : CustomQueryParam<'ctx, DateTimeOffset> =
+    CustomQueryParam<'ctx, DateTimeOffset>(queryParamName, fun _ -> parseDateTimeOffsetAllowMissingOffset >> Job.result)
 
   static member Enum(queryParamName, enumMap: (string * 'a) list) =
     let d = dict enumMap